requirements.txt
absl-py                 0.15.0
astunparse              1.6.3
cached-property         1.5.2
cachetools              5.3.0
certifi                 2023.5.7
charset-normalizer      3.1.0
cycler                  0.11.0
flatbuffers             1.12
fonttools               4.38.0
gast                    0.4.0
google-auth             2.18.1
google-auth-oauthlib    0.4.6
google-pasta            0.2.0
grpcio                  1.34.1
h5py                    3.1.0
idna                    3.4
importlib-metadata      6.6.0
keras-nightly           2.5.0.dev2021032900
Keras-Preprocessing     1.1.2
kiwisolver              1.4.4
Markdown                3.4.3
MarkupSafe              2.1.2
matplotlib              3.5.3
numpy                   1.19.5
oauthlib                3.2.2
opencv-python           4.7.0.72
opt-einsum              3.3.0
packaging               23.1
Pillow                  9.5.0
tensorboard-plugin-wit  1.8.1
tensorflow              2.5.0
tensorflow-estimator    2.5.0
termcolor               1.1.0
typing-extensions       3.7.4.3
urllib3                 1.26.15
Werkzeug                2.2.3
wheel                   0.40.0
wrapt                   1.12.1
zipp                    3.15.0
